% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_process_reporter.R
\name{auto_start_process_reporter}
\alias{auto_start_process_reporter}
\title{Auto-start Process Reporter if needed}
\usage{
auto_start_process_reporter(hostname = Sys.info()["nodename"], con = NULL)
}
\arguments{
\item{hostname}{Hostname to check/start reporter on}

\item{con}{Database connection}
}
\value{
TRUE if reporter is running (or started), FALSE if unable to start
}
\description{
Internal function called by task_start() to ensure a process reporter
is running when tasks are started.
}
\keyword{internal}
