% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_reporter_schema.R
\name{check_reporter_tables_exist}
\alias{check_reporter_tables_exist}
\title{Check if Reporter Tables Exist}
\usage{
check_reporter_tables_exist(conn, driver)
}
\arguments{
\item{conn}{Database connection}

\item{driver}{Database driver type}
}
\value{
TRUE if all required tables exist
}
\description{
Check if Reporter Tables Exist
}
\keyword{internal}
