% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_snippets}
\alias{copy_snippets}
\title{To copy snippets to computers}
\usage{
copy_snippets(from = file.path(system.file(package = "taskeR"), "snippets"),
  to = file.path(Sys.getenv("HOME"), ".R", "snippets"))
}
\description{
To copy snippets to computers
}
