% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_connection.R
\name{create_schema}
\alias{create_schema}
\title{Create tasker database schema}
\usage{
create_schema(conn = NULL, schema = NULL)
}
\arguments{
\item{conn}{Database connection (optional, will create if NULL)}

\item{schema}{Schema name (default: from config)}
}
\value{
TRUE on success
}
\description{
Create tasker database schema
}
\examples{
\dontrun{
create_schema()
}
}
