% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_process_reporter_schema.R
\name{drop_process_reporter_tables}
\alias{drop_process_reporter_tables}
\title{Drop Process Reporter Tables}
\usage{
drop_process_reporter_tables(conn, driver)
}
\arguments{
\item{conn}{Database connection}

\item{driver}{Database driver type}
}
\description{
Drop Process Reporter Tables
}
\keyword{internal}
