% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_reporter_schema.R
\name{drop_reporter_tables}
\alias{drop_reporter_tables}
\title{Check if Reporter Tables Exist}
\usage{
drop_reporter_tables(conn, driver)
}
\arguments{
\item{conn}{Database connection}

\item{driver}{Database driver type}
}
\value{
TRUE if all required tables exist
}
\description{
Check if Reporter Tables Exist
}
\keyword{internal}
