% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_env_vars.R
\name{expand_env_vars}
\alias{expand_env_vars}
\title{Expand environment variables in configuration}
\usage{
expand_env_vars(config)
}
\arguments{
\item{config}{Configuration list}
}
\value{
Configuration list with expanded variables
}
\description{
Expand environment variables in configuration
}
\keyword{internal}
