% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.task.R
\name{extract_info}
\alias{extract_info}
\title{To extract a specific property of a task}
\usage{
extract_info(object, info = c("proj_tags", "proj_start", "proj_end", "todo",
  "info", "tags", "personnel", "cost", "deadline", "priority", "start", "end"))
}
\arguments{
\item{object}{A task object}

\item{info}{One of pre-defined property of task to extract}
}
\value{
A character vector contains task's property to extract
}
\description{
To extract a specific property of a task
}
