% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.task.R
\name{extract_task}
\alias{extract_task}
\title{To extract all task from an .Rmd file}
\usage{
extract_task(file)
}
\arguments{
\item{file}{Path to an .Rmd file}
}
\value{
A 'task' object which is a character vector contains all task information extracted from the .Rmd file
}
\description{
To extract all task from an .Rmd file
}
