% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_config_file.R
\name{find_config_file}
\alias{find_config_file}
\title{Find .tasker.yml configuration file}
\usage{
find_config_file(start_dir = getwd(), filename = ".tasker.yml", max_depth = 10)
}
\arguments{
\item{start_dir}{Starting directory (default: current working directory)}

\item{filename}{Configuration filename (default: ".tasker.yml")}

\item{max_depth}{Maximum directory levels to search up (default: 10)}
}
\value{
Path to config file, or NULL if not found
}
\description{
Find .tasker.yml configuration file
}
