% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion_estimation.R
\name{format_duration_seconds}
\alias{format_duration_seconds}
\title{Format duration from seconds to human readable format}
\usage{
format_duration_seconds(seconds)
}
\arguments{
\item{seconds}{Number of seconds}
}
\value{
Character string with formatted duration
}
\description{
Format duration from seconds to human readable format
}
