% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_reporter_main_loop.R,
%   R/reporter_main_loop.R
\name{get_active_tasks_for_reporter}
\alias{get_active_tasks_for_reporter}
\title{Get active tasks for a hostname}
\usage{
get_active_tasks_for_reporter(con, hostname)

get_active_tasks_for_reporter(con, hostname)
}
\arguments{
\item{con}{Database connection}

\item{hostname}{Host to check for active tasks}
}
\value{
List of task info (run_id, process_id, etc.)

List of task info (run_id, process_id, etc.)
}
\description{
Internal function to retrieve tasks that are currently running
and need process monitoring.

Internal function to retrieve tasks that are currently running
and need process monitoring.
}
\keyword{internal}
