% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parent_id.R, R/utils.R
\name{get_parent_pid}
\alias{get_parent_pid}
\title{Get parent process ID}
\usage{
get_parent_pid()

get_parent_pid()
}
\value{
Parent PID or NULL

Parent PID or NULL
}
\description{
Get parent process ID

Get parent process ID
}
\keyword{internal}
