% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_task_status.R
\name{get_stages}
\alias{get_stages}
\title{Get all stages}
\usage{
get_stages(conn = NULL)
}
\arguments{
\item{conn}{Database connection (optional)}
}
\value{
Data frame with stages
}
\description{
Get all stages
}
