% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.task.R
\name{get_task}
\alias{get_task}
\title{To get all properties of all task from an .Rmd file and export into a data frame}
\usage{
get_task(file)
}
\arguments{
\item{file}{Path to an .Rmd file}
}
\value{
A data frame contains all properties of all task from the file
}
\description{
To get all properties of all task from an .Rmd file and export into a data frame
}
