% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_tasker_config}
\alias{get_tasker_config}
\title{Get current tasker configuration}
\usage{
get_tasker_config()
}
\value{
List with configuration settings, or NULL if not loaded
}
\description{
Get current tasker configuration
}
