% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{get_tasks}
\alias{get_tasks}
\title{Get registered tasks}
\usage{
get_tasks(stage = NULL, name = NULL, conn = NULL)
}
\arguments{
\item{stage}{Filter by stage (optional)}

\item{name}{Filter by task name (optional)}

\item{conn}{Database connection (optional)}
}
\value{
Data frame with task information
}
\description{
Get registered tasks
}
\examples{
\dontrun{
get_tasks()
get_tasks(stage = "PREREQ")
}
}
