% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_env_config}
\alias{load_env_config}
\title{Load configuration from environment variables}
\usage{
load_env_config()
}
\value{
List with configuration from environment
}
\description{
Load configuration from environment variables
}
\keyword{internal}
