% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.task.R
\name{plot_task}
\alias{plot_task}
\title{To plot a task.summary object}
\usage{
plot_task(summary_info)
}
\arguments{
\item{summary_info}{A task.summary object}
}
\value{
A plot created by ggplot
}
\description{
To plot a task.summary object
}
