% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_params}
\alias{prepare_params}
\title{Prepare parameters for SQL queries (handle NULLs)}
\usage{
prepare_params(...)
}
\arguments{
\item{...}{Parameters to prepare}
}
\value{
List of parameters with NULL converted to NA
}
\description{
Prepare parameters for SQL queries (handle NULLs)
}
\keyword{internal}
