% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter_status.R
\name{register_reporter}
\alias{register_reporter}
\title{Register or update reporter in database}
\usage{
register_reporter(
  con,
  hostname,
  process_id,
  version = as.character(packageVersion("tasker"))
)
}
\arguments{
\item{con}{Database connection}

\item{hostname}{Reporter hostname}

\item{process_id}{Reporter PID}

\item{version}{Package version}
}
\value{
TRUE on success
}
\description{
Internal function to register a reporter or update its registration.
Uses UPSERT to handle concurrent starts.
}
\keyword{internal}
