% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.task.R
\name{split_task}
\alias{split_task}
\title{To split task vector into a list of tasks}
\usage{
split_task(object)
}
\arguments{
\item{object}{A 'task' object which is a character vector contains all task information extracted from the .Rmd file}
}
\value{
A list of all tasks and their properties
}
\description{
To split task vector into a list of tasks
}
