% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.task.R
\name{summarize_task}
\alias{summarize_task}
\title{To summarize tasks from a data frame of all tasks}
\usage{
summarize_task(task_info)
}
\arguments{
\item{task_info}{A task.frame object}
}
\value{
A data frame contains all property and additional summaries of all tasks
}
\description{
To summarize tasks from a data frame of all tasks
}
