% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{tasker-hooks}
\alias{tasker-hooks}
\alias{.onLoad}
\alias{.onAttach}
\title{Package loading hooks}
\usage{
.onLoad(libname, pkgname)

.onAttach(libname, pkgname)
}
\description{
Automatically configures logger with sensible defaults when tasker is loaded.
}
\keyword{internal}
