% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_reporter_schema.R
\name{drop_reporter_tables}
\alias{drop_reporter_tables}
\title{Drop Reporter Tables}
\usage{
drop_reporter_tables(conn, driver)
}
\arguments{
\item{conn}{Database connection}

\item{driver}{Database driver type}
}
\description{
Drop Reporter Tables
}
\keyword{internal}
