% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasker_config.R, R/utils.R
\name{ensure_configured}
\alias{ensure_configured}
\title{Ensure configuration is loaded}
\usage{
ensure_configured()

ensure_configured()
}
\value{
TRUE if configured

TRUE if configured
}
\description{
Ensure configuration is loaded

Ensure configuration is loaded
}
\keyword{internal}
