% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_active_tasks.R,
%   R/process_reporter_main_loop.R
\name{get_active_tasks}
\alias{get_active_tasks}
\title{Get active (running) tasks}
\usage{
get_active_tasks(con, hostname)

get_active_tasks(con, hostname)
}
\arguments{
\item{con}{Database connection}

\item{hostname}{Host to check for active tasks}

\item{conn}{Database connection (optional)}
}
\value{
Data frame with active tasks

List of task info (run_id, process_id, etc.)
}
\description{
Internal function to retrieve tasks that are currently running
and need process monitoring.
}
\examples{
\dontrun{
get_active_tasks()
}
}
\seealso{
\code{\link[=get_task_status]{get_task_status()}} to query all task statuses with filtering,
\code{\link[=get_task_history]{get_task_history()}} to view execution history, \code{\link[=task_start]{task_start()}} to start
tracking a task
}
\keyword{internal}
