% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_connection.R
\name{get_db_connection}
\alias{get_db_connection}
\title{Get database connection}
\usage{
get_db_connection()
}
\value{
DBI connection object
}
\description{
Get database connection
}
\examples{
\dontrun{
conn <- get_db_connection()
DBI::dbDisconnect(conn)
}
}
