% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reporter_database_status.R
\name{get_reporter_database_status}
\alias{get_reporter_database_status}
\title{Get reporter database status}
\usage{
get_reporter_database_status(hostname = NULL, con = NULL)
}
\arguments{
\item{hostname}{Check specific host (NULL = current host)}

\item{con}{Database connection (NULL = get default)}
}
\value{
Data frame with reporter status or NULL if not running
}
\description{
Checks if a reporter is running on a specific host.
}
