% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_name}
\alias{get_table_name}
\title{Get table name with or without schema}
\usage{
get_table_name(table, conn)
}
\arguments{
\item{table}{Table name}

\item{conn}{Database connection}
}
\value{
SQL identifier with schema prefix if applicable
}
\description{
Get table name with or without schema
}
\keyword{internal}
