% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tasks.R
\name{get_tasks}
\alias{get_tasks}
\title{Get all tasks (alias for get_task_status)}
\usage{
get_tasks(...)
}
\arguments{
\item{...}{Arguments passed to get_task_status}
}
\value{
Data frame with task status
}
\description{
Get all tasks (alias for get_task_status)
}
\examples{
\dontrun{
get_tasks()
}
}
