% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reporter.R
\name{is_reporter_alive}
\alias{is_reporter_alive}
\title{Check if a reporter process is alive}
\usage{
is_reporter_alive(process_id, hostname)
}
\description{
Check if a reporter process is alive
}
\keyword{internal}
