% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasker_config.R
\name{load_yaml_config}
\alias{load_yaml_config}
\title{Load YAML configuration file}
\usage{
load_yaml_config(config_file)
}
\arguments{
\item{config_file}{Path to YAML file}
}
\value{
List with configuration
}
\description{
Load YAML configuration file
}
\keyword{internal}
