% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{merge_configs}
\alias{merge_configs}
\title{Merge two configuration lists}
\usage{
merge_configs(base, overlay)
}
\arguments{
\item{base}{Base configuration}

\item{overlay}{Configuration to overlay}
}
\value{
Merged configuration
}
\description{
Merge two configuration lists
}
\keyword{internal}
