% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_reporter_status.R
\name{update_reporter_heartbeat}
\alias{update_reporter_heartbeat}
\title{Update reporter heartbeat timestamp}
\usage{
update_reporter_heartbeat(con, hostname)
}
\arguments{
\item{con}{Database connection}

\item{hostname}{Reporter hostname}
}
\value{
Number of rows updated
}
\description{
Internal function to update the last_heartbeat timestamp for a reporter.
}
\keyword{internal}
