% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasker_config.R
\name{validate_config}
\alias{validate_config}
\title{Validate configuration}
\usage{
validate_config(config)
}
\arguments{
\item{config}{Configuration list}
}
\value{
TRUE if valid (or stops with error)
}
\description{
Validate configuration
}
\keyword{internal}
